requirements.txt
numpy==1.20.0
pylab-sdk==1.3.2
requests==2.25.1
scipy==1.6.0
matplotlib==3.3.4